+++ tcod/ecs/constants.py
@@ -2,9 +2,18 @@
 
 from __future__ import annotations
 
-from typing import Final
 
-from sentinel_value import sentinel
+class _IgnoreSetState(type):
+    def __setstate__(cls, _state: object) -> None:
+        """Ignore setstate on outdated sentinel-value pickle data."""
 
-IsA: Final = sentinel("IsA")
-"""The default is-a relationship tag used for entity inheritance."""
+
+class IsA(metaclass=_IgnoreSetState):
+    """The default is-a relationship tag used for entity inheritance."""
+
+    def __new__(cls: type[IsA], *_args: object) -> type[IsA]:  # type: ignore[misc]
+        """Return own type instead of instance, for outdated sentinel-value pickle data."""
+        return cls
+
+
+_sentinel_IsA = IsA  # Compatibility with sentinel-value, deprecated since 5.4  # noqa: N816