+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Removed dependency on `sentinel-value`.
+
 ## [5.4.2] - 2025-11-28
 
 ### Removed